binary and segment tree

+++ src/main/kotlin/com/falcinspire/hacklib/math/BitSetAsNumber.kt
@@ -3,4 +3,4 @@ package com.falcinspire.hacklib.math
 import java.util.*
 
 fun BitSet.asInt() = toLongArray()[0].toInt()
-fun BitSet.asLong() = toLongArray()[0].toLong()
+fun BitSet.asLong() = toLongArray()[0]
